fix: deadlock when replacing stream sink in SendToDartLogger
When setting a new stream sink in `SendToDartLogger`, we lock the global
rw lock for writing. However, if the lock already contained a sink, it
was droppen at the same line when replacing it. A stream sink logs a
message on drop, and therefore `SendToDartLogger` tries to lock the
global rw lock again, this time for reading. The lock is not reentrant,
so it deadlocks.

This makes sure we drop the previous sink *after* unlocking the rw lock.

Fixes #225