feat: introduce global navigation and routing
This introduces an application router `AppRouter` which listens to the
state provided by the `NavigationCubit` and builds the tree of pages
accordingly. In particular, all navigation is done via the
`NavigationCubit`. All pages are built inside the `AppRouter`.

The core client is moved below the `App` widget. From it we derive the
`UserState` which switches between the `Into` and the `Home` screens
depending on whether the user is logged in or not.

The registration flow is implemented by the `RegistrationCubit`. All
registration screens are now stateless widgets, which operate on this
cubit.

What is not done yet:

Widgets are still using the state from the `CoreClient` directly. In
particular, the current conversation is not taken from the navigation
state, but from the `CoreClient`.