refactor: expose the user via a provider in the whole app (#249)
* refactor: expose the user via a provider in the whole app

The `UserCubit` is accessible in the whole app, after a user is loaded.
All widgets that access the user information e.g. their user name or
profile, are doing it via the `UserCubit`. In particular, it makes them
stateless and they automatically re-rendered when the user information
is changed.

Implementation details:

The cubit is implemented on the Rust side, and
is exposed to Flutter via frb. On the Flutter side, the `UserCubit` is a
wrapper around the `UserCubit` in Rust. Currently, it is not possible to
require that a from-Rust generated class in Dart impements a specific
interface. When this will be possible, the wrapper can be removed. ([1])

Also:
* rename `ObservableUser` to `LoadableUserCubit`
* CI: don't flutter clean and pub get in `just frb-generate`

[1]: https://github.com/phnx-im/infra/issues/248