refactor: structured logging in applogic
Use tracing for logging in applogic. Replace the simplelog
implementation with tracing-subscriber. Implement a tracing layer for
sending logs to the Dart side. The layer is used in Android and iOS. On
desktop, the logs are printed to the console via the tracing-subscriber
formatter.

Also configure the logging formatting in Dart to use the same format as
the Rust tracing logging. Logs from Dart are prefixed with `[F]` and
logs from Rust are prefixed with `[R]`.