feat: add conversation list cubit
The conversation list cubit contains the logic for loading
conversations. Its state is a list of conversations. The conversation
list view is now a stateless widget that uses this cubit.

Note: A conversation is not automatically opened on desktop now. Most
likely the currently opened conversation should be stored in the local
storage. Automatically opening a conversation is not consistent between
the desktop and mobile clients.

Also fix: too eager caching of conversation avatars.

A common implementation with a background emitter loop task was
extracted in to the `CubitCore` struct.