fix: flutter dev tools performance and inspector (#281)
When toggling flutter the performance overview or the widget inspector
in flutters dev tools, the root providers were recreated, which
triggered an exception:

```
Error: Could not find the correct Provider<UserCubit> above this ConversationViewContainer Widget
```

This was due to the fact that the `UserCubit` was not available for a
short time after the `LoadableUserCubit` was created. Moving providers
before the `MaterialApp` solved the issue.