feat: add message list and message cubits (#282)
The `MessageListCubit` is responsible for loading the last 500 messages
for a specific conversation. It allows retrieving a message by its index
and mapping a message ID to its index. On receiving incoming messages,
the list is reloaded, and the index lookup table is rebuilt. Currently,
it does not load additional messages when there are more than 500
messages in the conversation. It also calculates the newly introduced
`FlightPosition` of a message from it and its previous and next message
in the timeline. With a flight position each message can independently
decide how to be represented in the UI.

The `MessageCubit` is initialized with an already loaded message. It
listens to changes in the message and updates it accordingly. Whenever
the message is added or updated, it reloads and also recalculates its
flight position from the previous and next message in the timeline.

The `MessageViewList` now uses the `ListView.custom` constructor instead
of `ListView.builder`. The former makes it possible to reuse list tiles
based on their keys. In our case, we assign the message ID as the key.
When the list is rebuilt, a reverse lookup from the key (i.e., message
ID) to the index is performed. This avoids rebuilding tiles that are
already built. In particular, when a new message is sent, only the tile
for the new message is built.

The visibility detection of messages is now done via
https://pub.dev/packages/visibility_detector, which allows detecting the
visibility of individual messages. The `mark_as_read` method was added
to the `ConversationDetailCubit`. The method is new debounced via a
tokio watch channel and additional state in the cubit. 

Missing features from this port:

- Scrolling to the end when a new message arrives  
- Storing the message draft per conversation  
- Loading batches when there are more than 500 messages  (#287)