chore: add snaphot tests in CI
Tests are added for the following widgets:

- `HomeScreen` (desktop layout)
- `ConverstationList`
- `ConversationListContent`
- `Conversation`
- `MessageList`

By default, snapshots are created with loaded fonts, on physical screen
size of a Pixel 8 and pixel density of 3. The snapshot tests tolerance
is 0.022 for ignoring the differences in rendering on different
platforms. The snapshots are stored in git LFS. The tests are run in CI
on macOS and Linux runners.

The line height of a message text is now set to default, because
otherwise it creates inconsistent rendering on different platforms.