feat: select user on into screen (#312)
Rework developer settings and add "Change User" and "Log Out" actions.

In the "Change User" action, a selection dialog displays all available
users. When a user is changed, it is remembered as the default user,
meaning it will be used as the login user at the next app start.

Additionally, add an action to delete only the currently logged-in user
database.

Closes #310