feat: communicate store changes from background processing
A background processing task fetches new messages and stores them in the
database. This task is triggered by push notifications.

This commit introduces a store notifications queue. The background task
stores in the queue any changes made to the store in the background.
When the foreground application reconnects to the websocket, it dequeues
the pending store notifications and communicates them to the UI so it
can update its state.

Fixes #337