feat: QS Client <-> Server API Versioning (#325)
* feat: QS Client <-> Server API Versioning

Introduce a versioning scheme for the QS API in both directions.

API requests and responses are now tagged with a version (currently
Alpha = 1). If the version number is not supported, deserialization on
the server/client side will convert the request into the variant
`Other(version)`. The unsupported version and the list of supported
versions are returned as an error.

Errors from the server to the client are not versioned and are returned
as plain text.

Also, test all QS API endpoints, including requests and responses.