feat: implement API version negotiation for QS API
If the QS backend does not support the API version used by the client,
it responds with a `406 Not Acceptable` status and includes a header,
`x-accepted-api-versions`, which contains a comma-separated list of
supported API versions.

The client handles the `406` status code by attempting to resend the
request using the highest supported API version, if possible.

The latest supported API version is stored in the client's memory.