feat: collect logs locally
This commit changes the Rust logging setup to additionally collect logs
in a local file. The logs from the application and background execution
are split between two files: `app.log` and `background.log`. Both files
are stored in the application's cache directory. Currently, only logs
from Rust and Flutter are stored in the files.

Each file has a constant size of 500 KiB. It is memory-mapped and used
as a ring buffer. A zero byte (`\0`) is used as a delimiter between the
end and the beginning of the logs.

Additionally, a log viewer has been added to the developer settings
screen. The logs are displayed in reverse chronological order.