feat: share or save logs (#359)
Add a share and save menu item to the logs screen (in developer
settings). The share menu is available on Android and iOS, the save menu
is available on Linux, MacOS and Windows.

The logs are compressed as a tar gzip archive containing the
logs/app.log and logs/background.log files.

Closes #353