feat: introduce protobuf/grpc API
A protobuf server/client API spec is introduced in `protos/api`. The API
is not yet final but resembles the one currently implemented in the HTTP
server/backend. The crate also contains a build script to generate the
model and Tonic gRPC client/server code.

A new listener has been added to the server to accept gRPC connections.
The server currently serves the stub DS API.