feat(!): replace usernames by client id (#472)
This commit replaces `QualifiedUserName` with `AsClientId` across the
entire codebase. Users are now uniquely identified on AS by a UUID and a
domain.

## Database representation

In the database (both backend and frontend), a client ID is stored in
two separate columns: UUID and domain. This allows referencing users by
UUID alone, omitting the domain.

Additionally, conversation type and status are now stored in dedicated
columns to avoid serialization/deserialization. Past members are now
stored in a separate table.

## App

Removing usernames from the app also removes the registration screen for
choosing a username. Now, the user only needs to choose a domain, a
display name, and optionally an avatar.

## Misc

* The user profile key is now derived from the user ID.
* User connection is established via a UUID. For that, UUID is exposed in the
  user settings screen in the App.
* User profile of user if non-optional anymore. In case, it cannot be loaded
  or is unknown, a fallback is done via a profile generated from UUID.

## Follow-ups

Display names need to be propagated to the app. Previously, usernames
were also used in conversation titles and as message senders. Now, the
corresponding display name must be resolved dynamically based on the
client ID. TODOs were added in the code where this needs to happen.

`AsClientId` is effectively a user ID, so the type should be renamed to
`UserId`. All mentions of "client ID" should be updated to "user ID."