feat: add localization support
This commit adds support for localizing the app's UI text.  It
introduces the `intl` package and generates localized strings for
English. All relevant UI text has been updated to use localized strings.

The generated localization files are checked in to the repository. CI
checks that the localization files are up to date.