feat: resolve user profiles and display names uniformly
A new `ContactsCubit` is introduced to manage user profiles. It caches
already loaded user profiles and provides a synchronous API to get a
user profile, a display name, or a profile picture. In case, the user
profile is not yet in cache, it is loaded in the background. The state
of the cubit is opaque and uses clone-on-write semantics.

The logic for loading own user profile and user profiles of contacts is
moved from the `UserCubit` to the `ContactsCubit`.

Closes #475