feat: add delete and leave conversations
This commit adds and implements delete and delete/leave conversation
buttons to the details of a connection resp. group.

Also fixed the database schema to ensure that deleting a conversation
cascades to the related data. Also fixed an invalid trigger which was
referencing a non-existent field in the `indexed_keys` table.

Closes #510, closes #511