feat: add attachments backend API
This commit adds support for attachments to the delivery service.

Attachments are provisioned using a new `ProvisionAttachment` gRPC
method. This method returns a pre-signed URL for uploading the
attachment, along with the necessary headers. A new `GetAttachmentUrl`
method retrieves a pre-signed URL for downloading attachments.

Minio is added to docker compose as a development storage backend.