feat: implement uploading of attachments (#586)
This commit implements the functionality for uploading attachments and
sending messages that include them.

Attachments are encrypted with Aes256Gcm12 and hashed with Blake3.
Image attachments are resized to a maximum of 4096x4096 pixels
and encoded in the WebP format (latter subject to change). The blurhash
and dimensions of the image are stored in the message content.

This commit does not include the rendering of attachments in the app.