feat: introduce message drafts (#609)
This commit introduces message drafts to the application.

A message consists of a text and an optional edited message id
(currently not used) per conversation. Drafts are persisted in the
database.

A message draft consists of text and an optional editing message id
(currently unused) per conversation. Drafts are persisted in the
database.

Behavior:

* When typing a message, the draft is saved with a debounce of 500ms.
* When a conversation is opened, the draft is loaded asynchronously.
* The draft is shown in the list of conversations, except when the
  conversation is currently selected.
* Conversation are sorted first by the last updated draft, and then by
  the last message timestamp.

Other fixes:

* Show a clippy or image emoji for messages without text that contain a
  file or image attachment, respectively.
* Clean up more composer state on dispose.
* Use image picker instead of file picker for attachments upload.