refactor: context menu (#617)
Introduce a context menu anchor widget and a context menu controller.
The former wraps a child widget and shows a context menu when tapped.
The controller is used to show and hide the menu. Also it hides the menu
on Escape key press.

The menu is now rendered in two passes. First pass calculates the size
of the menu. Second pass renders the menu at the calculated position.
The scaling of the interface is taken into account.

The width of the menu is now the intrinsic width of the menu items, that
is, the width of the widest item.