feat: help and contact us pages (#651)
This commit adds a help screen to user settings. The help screen
contains **Contact Us**, **Version Info**, and **Licenses** buttons.

The **Contact Us** and **Version Info** buttons have been implemented.
The **Licenses** button opens the licenses screen, but it does not yet
display any licenses.

The **Contact Us** screen includes a subject dropdown and a body text
field. Both are validated on submission. If the validation is
successful, the app attempts to open the default email client with the
provided subject and body. If no email client is available, an error
message is displayed to the user.