feat: block and unblock contacts (#691)
This commit introduces the ability to **block and unblock contacts**,
along with UI changes to reflect this new functionality. The actual
blocking of messages and profile sharing hasn't been implemented yet.

Blocked contacts are stored in a separate database table. The status of
the corresponding one-on-one chat is set to `Blocked`. In the UI, a
contact can be blocked within the chat details (for one-on-one chats)
and unblocked from either the same location or the chat message view.
Instead of a message composer, we display buttons to delete the chat or
unblock the contact.