fix: stale listeners in qs queues (#706)
This commit fixes the issue of **stale listeners** remaining in the QS
queues. Consequently, **push notifications** weren't being sent to
clients that had previously listened to a queue.

On the **backend**, we now cancel the listening task when a client
closes its incoming request stream.

On the **client**, the connection to the server is now bound to the
lifetime of the stream returned by the `listen_queue` method. This
stream is dropped when the app enters the background. Previously, the
connection wasn't closed immediately.

Another unrelated bug is fixed: When going into the background, the
background task **dropped the stream** and entered a **backoff state**.
Instead, it should enter the **initial state** and wait for the app to
be in the foreground again.

This regression was introduced in #672.