chore: use widget state to determine button style (#719)
The button style is now fully defined in the theme data. The style of an
inactive button is now defined as part of the widget state mapping. In
particular, this avoids passing custom styles to buttons, and allows to
override the style via `copyWith`. Also, this fixes wrong styles for 
disabled buttons and unifies the style between all buttons.