fix: enqueue QS message race (#724)
Enqueuing a message into a QS queue involves three database operations:

1. Load the client record from the database.
2. Insert the message into the queue.
3. Update the queue ratchet in the database.

These operations were wrapped in a transaction; however, the default
transaction isolation level in Postgres (read committed) does not
prevent the following races.

A race between steps 1 and 2 is detected by comparing the counters
stored in the queue ratchet and the queue table. When this occurs, a
message is lost, but the state remains consistent.

The race between steps 2 and 3, however, is not detected. In this case,
a queue ratchet may be overwritten by a concurrent enqueue operation,
which corrupts the state.

This commit makes sure that the client record is locked for the duration
of the transaction.

---------

Co-authored-by: Konrad Kohbrok <konrad.kohbrok@datashrine.de>