refactor: rely on QS ratchet sequence number only
Currently, we have the QS message number stored in the QS ratchet
counter of the client record, and in the queue data table. The source of
truth is the ratchet and since we don't use the counter in the queue
data tabel, we can safely remove it.

Also remove the QueueNotifier trait, which is used only once in the same
submodule.